Refactor DeleteCaseCommandTest into Unit and Integration Testing
