+++ src/main/java/seedu/pivot/logic/commands/casecommands/DeleteCaseCommand.java
@@ -26,7 +26,13 @@ public class DeleteCaseCommand extends DeleteCommand {
 
     private final Index targetIndex;
 
+    /**
+     * Creates an DeleteCaseCommand to delete the {@code Case} specified at {@code Index}.
+     *
+     * @param targetIndex Index of Case to be deleted from PIVOT.
+     */
     public DeleteCaseCommand(Index targetIndex) {
+        requireNonNull(targetIndex);
         this.targetIndex = targetIndex;
     }
