+++ src/test/java/seedu/pivot/logic/commands/AddCaseCommandTest.java
@@ -6,21 +6,16 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.pivot.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.pivot.commons.core.GuiSettings;
 import seedu.pivot.logic.commands.casecommands.AddCaseCommand;
 import seedu.pivot.logic.commands.exceptions.CommandException;
-import seedu.pivot.model.Model;
+import seedu.pivot.logic.commands.testutil.ModelStub;
 import seedu.pivot.model.Pivot;
 import seedu.pivot.model.ReadOnlyPivot;
-import seedu.pivot.model.ReadOnlyUserPrefs;
 import seedu.pivot.model.investigationcase.Case;
 import seedu.pivot.testutil.CaseBuilder;
 
@@ -75,81 +70,6 @@ public class AddCaseCommandTest {
 
         // different person -> returns false
         assertFalse(addAliceCommand.equals(addBobCommand));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getPivotFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPivotFilePath(Path pivotFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addCase(Case investigationCase) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPivotBook(ReadOnlyPivot newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyPivot getPivot() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasCase(Case investigationCase) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteCase(Case target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setCase(Case target, Case editedCase) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Case> getFilteredCaseList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredCaseList(Predicate<Case> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 
     /**
