+++ src/test/java/seedu/pivot/logic/commands/ClearCommandTest.java
@@ -1,6 +1,6 @@
 package seedu.pivot.logic.commands;
 
-import static seedu.pivot.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static seedu.pivot.logic.commands.testutil.CommandTestUtil.assertCommandSuccess;
 import static seedu.pivot.testutil.TypicalCases.getTypicalPivot;
 
 import org.junit.jupiter.api.Test;
