+++ src/test/java/seedu/pivot/logic/commands/FindCommandTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.pivot.commons.core.Messages.MESSAGE_CASES_LISTED_OVERVIEW;
-import static seedu.pivot.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static seedu.pivot.logic.commands.testutil.CommandTestUtil.assertCommandSuccess;
 import static seedu.pivot.testutil.TypicalCases.CARL;
 import static seedu.pivot.testutil.TypicalCases.ELLE;
 import static seedu.pivot.testutil.TypicalCases.FIONA;
